feat(agent): add tool dedup, trust decay, snapshot/rollback, validation, small model mode, and model escalation
Implement six features from smallcode evaluation plan:

- Tool dedup cache: caches read-only tool results (read/grep/glob/index)
  keyed by hash(tool+args), bounded to 64 entries with FIFO eviction,
  cleared on compaction. Prefixed with [cached] on hits.

- Trust decay: tracks per-tool consecutive failures. After configurable
  thresholds (warn_after=3, drop_after=5), tools are demoted/dropped.
  min_tools safeguard prevents dropping below minimum. Configurable via
  [agent.trust_decay] in config.

- Snapshot & rollback: auto-snapshots files before write/edit/multiedit
  operations. Commits on agent Done, rollback via /undo command.
  Thread-safe via Arc<Mutex>. Files >5MB and outside workdir skipped.

- Post-write validation: detects project type (Rust/TS/Go/Python) from
  config files and runs validation commands after writes. Disabled by
  default, configurable via [agent.validation].

- Small model mode: auto-detects models with context_window < 32k.
  Reduces tools to core set, uses compact system prompt, triggers
  compaction at 50% instead of 80%, adds aggressive JSON repair on
  parse failures. Configurable via [agent.small_model].

- Model escalation: tracks per-model failure rates, emits
  ModelEscalation event when rate >= 60% after 5 calls. UI handles
  event with Action::ChangeModel for automatic tier upgrade. Tier
  inferred from model ID (haiku/flash→Weak, sonnet→Default, opus→Strong).

New modules: dedup, trust, snapshot, validation, escalation.
New config types: TrustDecayConfig, ValidationConfig, SmallModelConfig.
New event: AgentEvent::ModelEscalation.
New command: ExtractedCommand::Undo.

45 tests added across all new modules.

+++ craft-agent/src/agent/compaction.rs
@@ -245,9 +245,9 @@ pub(super) fn progressive_compact(
 
 /// Check if estimated history tokens have reached the proactive compression threshold.
 /// This fires before overflow to compress incrementally rather than all-at-once.
-pub(super) fn is_proactive_threshold(history: &History, model: &Model, ratio: f32) -> bool {
+pub(super) fn is_proactive_threshold(history: &History, model: &Model, ratio: f64) -> bool {
     let estimated = history.estimate_tokens(model);
-    let threshold = (model.context_window as f32 * ratio) as u32;
+    let threshold = (model.context_window as f64 * ratio) as u32;
     estimated >= threshold
 }
